add BMSTree class to manage bms folder add save, load function (test not completed) modify function to create BMSInfoData instance in heap memory fix `player` header only to confirm KeyType and not to save fix to save only the file name instead of the entire path in BMSInfoData remove implicit casts

+++ libTest/Serializer.h
@@ -30,7 +30,7 @@ namespace bms {
 	template<class T>
 	inline void WriteToBinaryImpl(std::ostream& os, const typename std::vector<T>& v) {
 		// store count of vector element
-		WriteToBinary<uint32_t>(os, v.size());
+		WriteToBinary<uint32_t>(os, static_cast<uint32_t>(v.size()));
 		for (auto& e : v) {
 			WriteToBinary(os, e);
 		}
@@ -38,7 +38,7 @@ namespace bms {
 	template<class T>
 	inline void WriteToBinaryImpl(std::ostream& os, const typename std::basic_string<T>& v) {
 		// store count of char
-		WriteToBinary<uint32_t>(os, v.size());
+		WriteToBinary<uint32_t>(os, static_cast<uint32_t>(v.size()));
 		for (auto& e : v) {
 			WriteToBinary(os, e);
 		}
@@ -51,7 +51,7 @@ namespace bms {
 	template<class T1, class T2>
 	inline void WriteToBinaryImpl(std::ostream& os, const typename std::unordered_map<T1, T2>& v) {
 		// store count of unordered_map element
-		WriteToBinary<uint32_t>(os, v.size());
+		WriteToBinary<uint32_t>(os, static_cast<uint32_t>(v.size()));
 		for (auto& e : v) {
 			WriteToBinary(os, e);
 		}
